+++ node/test/go/dev-mode/test/integration/index.test.js
@@ -70,6 +70,40 @@ describe('Integration', function () {
       },
     ],
     [
+      'timeout',
+      {
+        variants: new Map([
+          [
+            'v16',
+            {
+              configuration: { Runtime: 'nodejs16.x', Timeout: 3 },
+              expectedOutcome: 'error:unhandled',
+              isCustomResponse: true,
+            },
+          ],
+          [
+            'v18',
+            {
+              configuration: { Runtime: 'nodejs18.x', Timeout: 3 },
+              expectedOutcome: 'error:unhandled',
+              isCustomResponse: true,
+            },
+          ],
+        ]),
+        config: {
+          test: ({ invocationsData }) => {
+            for (const { reqResPayloads } of invocationsData) {
+              expect(reqResPayloads.length).to.equal(2);
+              const lastPayload = reqResPayloads[1];
+              const response = lastPayload.payload.requestResponse;
+              expect(response.tags.error.message).to.contain('Task timed out');
+              expect(response.tags.error.type).to.equal(1);
+            }
+          },
+        },
+      },
+    ],
+    [
       'with-internal',
       {
         variants: new Map([
@@ -128,6 +162,40 @@ describe('Integration', function () {
         },
       },
     ],
+    [
+      'timeout-with-internal',
+      {
+        variants: new Map([
+          [
+            'v18',
+            {
+              ...internalConfiguration,
+              configuration: {
+                ...internalConfiguration.configuration,
+                Runtime: 'nodejs18.x',
+                Timeout: 3,
+              },
+              expectedOutcome: 'error:unhandled',
+              isCustomResponse: true,
+            },
+          ],
+        ]),
+        config: {
+          test: ({ invocationsData }) => {
+            for (const { reqResPayloads, devModePayloads } of invocationsData) {
+              // TraceId should match the one send via the dev mode payload
+              const traceId = devModePayloads[0].payload?.trace.spans[0].traceId.toString();
+              expect(reqResPayloads.length).to.equal(2);
+              const lastPayload = reqResPayloads[1];
+              const response = lastPayload.payload.requestResponse;
+              expect(response.tags.error.message).to.contain('Task timed out');
+              expect(response.tags.error.type).to.equal(1);
+              expect(response.traceId.toString()).to.equal(traceId);
+            }
+          },
+        },
+      },
+    ],
   ]);
 
   const testVariantsConfig = resolveTestVariantsConfig(useCasesConfig);
@@ -172,8 +240,16 @@ describe('Integration', function () {
       log.debug('%s test result: %o', testConfig.name, testResult);
       const { expectedOutcome } = testConfig;
       const { invocationsData } = testResult;
-      if (expectedOutcome === 'success' || expectedOutcome === 'error:handled') {
-        if (expectedOutcome === 'success' && !testConfig.isAsyncInvocation) {
+      if (
+        expectedOutcome === 'success' ||
+        expectedOutcome === 'error:handled' ||
+        expectedOutcome === 'error:unhandled'
+      ) {
+        if (
+          expectedOutcome === 'success' &&
+          !testConfig.isAsyncInvocation &&
+          !testConfig.isCustomResponse
+        ) {
           for (const { responsePayload } of invocationsData) {
             expect(responsePayload.raw).to.equal('"ok"');
           }